% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_data.R
\name{archive_nwsldata}
\alias{archive_nwsldata}
\title{Archive nwsldata}
\usage{
archive_nwsldata(release_name)
}
\arguments{
\item{release_name}{Name of the release}
}
\description{
Archive nwsldata
}
\note{
Mod'd from Tan Ho
}
